requirements.txt
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt5Designer==5.14.1
mysql-connector-python==8.3.0